requirements.txt
Django==1.8.4
django-webpack-loader==0.1.3
djangorestframework==3.2.3
ipython==4.0.0
factory-boy==2.5.2
django-filter==0.11.0
